MGMT-5416 Fix outdated preflight HW requirements
The preflight cluster HW requirements are queried repeatedly to secure
correct data set to render.
