+++ src/components/AddHosts/InventoryAddHost.tsx
@@ -5,24 +5,9 @@ import { DiscoveryImageModalButton } from '../clusterConfiguration/discoveryImag
 import { DiscoveryTroubleshootingModal } from '../clusterConfiguration/DiscoveryTroubleshootingModal';
 import InformationAndAlerts from '../clusterConfiguration/InformationAndAlerts';
 import DiscoveryInstructions from '../clusterConfiguration/DiscoveryInstructions';
-import { fileSize } from '../hosts/utils';
 import { AddHostsContext } from './AddHostsContext';
-import { useClusterPreflightRequirementsContext } from '../clusterConfiguration/ClusterPreflightRequirementsContext';
+import { AddHostRequirementsContent } from '../hosts/HostRequirementsContent';
 import { isSingleNodeCluster } from '../clusters/utils';
-
-const HostRequirementsContent = () => {
-  const { preflightRequirements } = useClusterPreflightRequirementsContext();
-
-  const worker = preflightRequirements?.ocp?.worker?.quantitative;
-  const workerRam = fileSize((worker?.ramMib || 8 * 1024) * 1024 * 1024, 2, 'iec');
-
-  return (
-    <Text component="p">
-      Worker hosts must have at least {worker?.cpuCores || 2} CPU cores, {workerRam} of RAM, and{' '}
-      {worker?.diskSizeGb || 120} GB of filesystem storage.
-    </Text>
-  );
-};
 
 const InventoryAddHosts: React.FC = () => {
   const { cluster } = React.useContext(AddHostsContext);
@@ -45,7 +30,7 @@ const InventoryAddHosts: React.FC = () => {
         </Text>
         <InformationAndAlerts
           cluster={cluster}
-          HostRequirementsContent={HostRequirementsContent}
+          HostRequirementsContent={AddHostRequirementsContent}
           setDiscoveryHintModalOpen={setDiscoveryHintModalOpen}
         />
       </TextContent>
